// src/app/page/master/pegawai-struktur-gaji-by-golongan/pegawai-struktur-gaji-by-golongan.interface.ts
import { EventEmitter } from '@angular/core';

export interface  PegawaiStrukturGajiByGolongan {
    kode?: any;
    kdGolonganPegawai?: number;
    hargaSatuan?: string
    kdOperatorRate?: string;
    kdKomponenHarga?: string;
    status1?: string;
    noRec?: string;
    label?: any;
    value?: any;
    namaStatus?: string;
    id?: any;
}